Move SafelistStoreSpy to its own file
